+++ Sources/BraintreePayPal/BTPayPalRequestUserAction.swift
@@ -9,7 +9,7 @@
 ///  Setting the `BTPayPalVaultRequest.userAction` to `.setupNow` changes the button text to "Setup Now", conveying to
 ///  the user that the funding instrument will be set up for future payments.
 @objc public enum BTPayPalRequestUserAction: Int {
-    /// Default
+    @available(*, deprecated, message: "Use a specific user action like `.payNow`, `.continue`, or `.setupNow`. `.none` will be removed.")
     case none
 
     /// Pay Now - this value can only be passed for the `BTPayPalCheckoutRequest`
@@ -17,6 +17,12 @@
     
     /// Setup Now - this value can only be passed for the `BTPayPalVaultRequest`
     case setupNow
+
+    /// Continue
+    case `continue`
+
+    /// Reserved for unknown or unsupported values.
+    case unknown
 
     var stringValue: String {
         switch self {
@@ -26,6 +32,19 @@
             return "setup_now"
         default:
             return ""
+        }
+    }
+
+    var analyticsValue: String {
+        switch self {
+        case .payNow:
+            return "pay"
+        case .setupNow:
+            return "setup_now"
+        case .continue, .none:
+            return "continue"
+        case .unknown:
+            return "none"
         }
     }
 }